Break up Item Module
